Extract findMethod handler into commands/findMethodInClass
No behavior change: moves the anonymous gemstone.findMethod handler
out of extension.ts into an exported, unit-testable function. Adds
characterization tests for the paths that aren't changing, ahead of
the actual fix
